Refactor code in fingerprint scan blocks into separate source file.
